Implement Provider with react-redux
