+++ src/index.js
@@ -1,40 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
+import { Provider } from 'react-redux';
 import App from './components/App';
+import booksReducer from './reducers/books';
 
-const initialState = {
-  books: [
-    {
-      title: 'Functional Programming',
-      author: 'Luis Atencio',
-      category: 'javascript',
-      id: Math.floor(Math.random * 1040),
-    },
-    {
-      title: 'Ruby by Example: Concepts and Code',
-      author: 'Kevin C. Baird',
-      category: 'ruby',
-      id: Math.floor(Math.random * 1040),
-    },
-    {
-      title: 'Haskell Programming from First Principles',
-      author: 'Chris Allen, Julie Moronuki',
-      category: 'haskell',
-      id: Math.floor(Math.random * 1040),
-    },
-    {
-      title: 'Get Programming with Go',
-      author: 'Nathan Youngman, Roger Peppé',
-      category: 'go',
-      id: Math.floor(Math.random * 1040),
-    },
-  ],
-};
+const store = createStore(booksReducer);
 
 ReactDOM.render(
-  <React.StrictMode>
+  <Provider store={store}>
     <App />
-  </React.StrictMode>,
+  </Provider>,
   document.getElementById('root'),
 );
