+++ src/reducers/books.js
@@ -1,4 +1,6 @@
-const booksReducer = (state = [], action) => {
+import initialState from './index';
+
+const booksReducer = (state = initialState.books, action) => {
   switch (action.type) {
     case 'CREATE_BOOK':
       return [...state, [action.content]];
